rename packages, name files correctly
